backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dateparser==1.2.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==21.2.0
idna==3.10
numpy==2.2.5
packaging==25.0
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
waitress==3.0.2
whitenoise==6.8.2
# Additional dependencies for the project
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.24.1
kaleido==0.2.1
reportlab==4.2.2
openai==1.55.3
Pillow==11.1.0
dj-database-url==2.3.0
apify-client==1.7.1
pinecone-client==6.0.0
